Create queries.js and utils.js and move motheds from dom.js to them
