+++ src/js/dom.js
@@ -1,7 +1,10 @@
 import HtmlString from './htmlstring.js';
+import Quries from './queries.js';
+import Utils from './utils.js';
 
 class Dom {
   static get elements() {
+
     const {
       emptyListInfo,
       initialContent,
@@ -9,126 +12,25 @@ class Dom {
     } 
     = HtmlString.domStrings;
 
-    const appContainer = document.getElementById('soccer-nova');
+    const {
+      appContainer, 
+      addPlayerButton,
+      savePlayerButton, 
+      numberOfPlayers,
+      playersContainer, 
+      getAddPlayerForm, 
+    } 
+    = Quries.domQueries;
 
-    const addPlayerButton = () => document.getElementById('add-player');
-
-    const savePlayerButton = () => document.getElementById('save-player');
-
-    const numberOfPlayers = () => document.querySelectorAll('.player').length;
-
-    const playersContainer = () => document.getElementById('empty-list');
-
-    const getAddPlayerForm = () => document.querySelector('form');
-
-    const moveImageToLeft = (element) => {
-      const divWithPlayerImage = element.querySelector('.form-inner-1');
-      divWithPlayerImage.classList.add('player-inner-left');
-      divWithPlayerImage.classList.remove('form-inner-1');
-    }
-
-    const moveImageToRight = (element) => {
-      const divWithPlayerImage = element.querySelector('.form-inner-1');
-      divWithPlayerImage.classList.add('player-inner-right');
-      divWithPlayerImage.classList.remove('form-inner-1');
-    }
-
-    const  removeEmptyListInfoText = () => {
-      if (document.querySelector('.empty-list-info-text')) {
-        document.querySelector('.empty-list-info-text').remove();
-      }
-    }
-
-    const resetForm = () => {
-      const formInputs = [...getAddPlayerForm().querySelectorAll('input')];
-      formInputs.forEach((input) => { input.value = ''; });
-      formInputs[0].previousElementSibling.src = '';
-    };
-
-    const addPlayerToList = () => {
-
-      removeEmptyListInfoText();
-
-      const newPlayerDiv = document.createElement('div');
-      newPlayerDiv.classList.add('player');
-
-      const currentForm = getAddPlayerForm();
-      newPlayerDiv.innerHTML = currentForm.innerHTML;
-
-      const playerDeleteButton = newPlayerDiv.querySelector('.close-form');
-
-      if (numberOfPlayers() % 2 === 0) {
-
-        playerDeleteButton.className = 'delete-right';
-
-        moveImageToLeft(newPlayerDiv)
-
-      } else {
-
-        playerDeleteButton.className = 'delete-left';
-
-        moveImageToRight(newPlayerDiv)
-
-      }
-
-      const filledForm = [...newPlayerDiv.querySelectorAll('input')];
-      filledForm.forEach((input, index) => {
-
-        if (input?.files) {
-
-          const playerImage = input.previousElementSibling;
-          playerImage.classList.add('player-image');
-
-        } else {
-
-          input.parentElement.classList.add('player-info-text');
-          input.parentElement.innerHTML = `
-                        ${input.getAttribute('placeholder')}:  
-                        <strong class="info-txt">
-                            ${currentForm.querySelectorAll('input')[index].value.trim()} 
-                        </strong>
-                    `;
-                    
-        }
-
-      });
-
-      currentForm.insertAdjacentElement('beforebegin', newPlayerDiv);
-    };
-
-    const formIsValid = () => {
-      const filledForm = [...getAddPlayerForm().querySelectorAll('input')];
-
-      return filledForm.every((input) => input.value.trim().length > 1);
-    };
-
-    const heighLightErrors = () => {
-      const filledForm = [...getAddPlayerForm().querySelectorAll('input')];
-
-      filledForm.forEach( input => {
-        input.value.trim() < 2 ? input.classList.add('border-red') : null;
-      });
-
-      setTimeout( () => removeHeightLight(), 2000)
-    }
-
-    const removeHeightLight = () => {
-      const heighLightedForm = [...getAddPlayerForm().querySelectorAll('input')];
-      heighLightedForm.forEach( input =>  input.classList.remove('border-red') )
-    }
-
-    const replaceButton = (targetButton, identifier, text) => {
-      targetButton.id = identifier;
-      targetButton.innerHTML = text;
-    };
-
-    const displayImagePreview = (fileInputChange) => {
-      const src = URL.createObjectURL(fileInputChange.target.files[0]);
-
-      const previewImage = fileInputChange.target.previousElementSibling;
-
-      previewImage.src = src;
-    };
+    const {
+      resetForm,
+      addPlayerToList,
+      formIsValid,
+      heighLightErrors,
+      replaceButton,
+      displayImagePreview,
+    } 
+    = Utils.domFunctions( getAddPlayerForm, numberOfPlayers );
 
     return {
       appContainer,
